Move `LeaseEventArgs`
