+++ src/Ocelot/LoadBalancer/LeaseEventArgs.cs
@@ -1,0 +1,17 @@
+﻿using Ocelot.Values;
+
+namespace Ocelot.LoadBalancer;
+
+public class LeaseEventArgs : EventArgs
+{
+    public LeaseEventArgs(Lease lease, Service service, int serviceIndex)
+    {
+        Lease = lease;
+        Service = service;
+        ServiceIndex = serviceIndex;
+    }
+
+    public Lease Lease { get; }
+    public Service Service { get; }
+    public int ServiceIndex { get; }
+}
