+++ src/Ocelot/LoadBalancer/LoadBalancers/RoundRobin.cs
@@ -117,17 +117,3 @@ public class RoundRobin : ILoadBalancer
         _leasing.AddRange(newLeases);
     }
 }
-
-public class LeaseEventArgs : EventArgs
-{
-    public LeaseEventArgs(Lease lease, Service service, int serviceIndex)
-    {
-        Lease = lease;
-        Service = service;
-        ServiceIndex = serviceIndex;
-    }
-
-    public Lease Lease { get; }
-    public Service Service { get; }
-    public int ServiceIndex { get; }
-}
